+++ src/Components/qual-page/QualitativePage.scss
@@ -1,17 +1,32 @@
-.background {
-    padding-left: 16px;
-    padding-right: 16px;    
+.qualitative-page {
+	padding-left: 16px;
+	padding-right: 16px;
+
+	.category-buttons {
+		display: flex;
+		flex-direction: row;
+		flex-wrap: wrap;
+		justify-content: center;
+		column-gap: 8px;
+		row-gap: 8px;
+	}
+
+
+	.submitted-notes {
+		display: flex;
+		flex-direction: column;
+		align-items: flex-start;
+		row-gap: 8px;
+		text-align: left;
+
+		li {
+			p {
+				margin: 0;
+				padding-top: 4px;
+				padding-left: 8px;
+			}
+		}
+	}
 }
 
 
-.grid {
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    grid-gap: 10px;
-}
-
-
-.list {
-    display: grid;
-    place-items: left;
-}
